requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.4.0
piexif==1.1.3
werkzeug==3.0.1
